helpers for robust scaling
prevents an extreme margin from drowning out the color scale on the no-undock trading table
instead of the margin, e.g. `103348686.74%`, it will now show in grey `+1000%`

This will rarely trigger, though for the last ~12 hours the highest buy for Morphite in Hek/Nakugard has been 0.03, which triggered that result.